"improve" Generate.aid
The length of the function is the same, though with more efficient code. The description tag has also been improved (though with increased length)

Tested to increase speed, though a very negligible amount (approx. 0.00003 seconds per operation)

Auto-converts the input to the proper format. Not very important to chlib as-is, though it does add fault tolerance